+++ app/src/main/java/co/infinum/productive/helpers/ProjectTitleComparator.java
@@ -7,7 +7,7 @@ import co.infinum.productive.models.ProjectTile;
 /**
  * Created by mjurinic on 18.11.15..
  */
-public class ProjectTileComparator implements Comparator<ProjectTile> {
+public class ProjectTitleComparator implements Comparator<ProjectTile> {
 
     @Override
     public int compare(ProjectTile lhs, ProjectTile rhs) {
