+++ app/src/main/java/co/infinum/productive/listeners/OnTasksClickListener.java
@@ -1,0 +1,10 @@
+package co.infinum.productive.listeners;
+
+/**
+ * Created by noxqs on 12.12.15..
+ */
+public interface OnTasksClickListener {
+
+    void onTasksClick(int position);
+
+}
